LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-starting-style-expected.txt


FAIL Style rules within @starting-style are scoped assert_equals: expected "150px" but got "200px"
